requirements_dev.txt
asgiref==3.6.0
astroid==2.15.0
coverage==7.2.2
dill==0.3.6
Django==4.1.7
django-cors-headers==3.13.0
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
Pillow==9.4.0
platformdirs==3.1.1
PyJWT==2.6.0
pylint==2.17.0
pytz==2022.7.1
sqlparse==0.4.4
tomlkit==0.11.6
wrapt==1.15.0
PyMySQL==1.0.3
mysql-connector-python==8.0.33